backend/requirements.txt
Flask
Flask-Cors
pymongo
python-dotenv
opencv-python-headless
requests
